requirements.txt
boto3
pillow
requests
pytz
pandas
matplotlib
mplfinance